% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_sum.R
\name{lm_sum}
\alias{lm_sum}
\title{Linear Regression}
\usage{
lm_sum(y, x)
}
\arguments{
\item{y}{dependent variable vector "y"}

\item{x}{independent variable matrix "x"}
}
\value{
lm summary
}
\description{
Provides a Summary Output of Regression Model
}
